+++ src/ZfSnapEventDebugger/Module.php
@@ -2,16 +2,10 @@
 
 namespace ZfSnapEventDebugger;
 
-use Zend\EventManager\EventInterface;
-use Zend\EventManager\EventManager;
-use Zend\EventManager\EventsCapableInterface;
-use Zend\EventManager\SharedEventManagerInterface;
 use Zend\ModuleManager\Feature\AutoloaderProviderInterface;
 use Zend\ModuleManager\Feature\ConfigProviderInterface;
 use Zend\ModuleManager\Feature\InitProviderInterface;
 use Zend\ModuleManager\ModuleManagerInterface;
-use Zend\Stdlib\CallbackHandler;
-use Zend\Stdlib\PriorityQueue;
 
 /**
  * Module
@@ -21,14 +15,9 @@ use Zend\Stdlib\PriorityQueue;
 class Module implements InitProviderInterface, ConfigProviderInterface, AutoloaderProviderInterface
 {
     /**
-     * @var array
+     * @var TriggerEventListener
      */
-    protected $events = array();
-
-    /**
-     * @var string|null
-     */
-    protected $getcwd = null;
+    protected $listener;
 
     /**
      * @param ModuleManagerInterface $manager
@@ -36,132 +25,19 @@ class Module implements InitProviderInterface, ConfigProviderInterface, Autoload
     public function init(ModuleManagerInterface $manager)
     {
         $sharedManager = $manager->getEventManager()->getSharedManager();
-        $self = $this;
-        $sharedManager->attach('*', '*', function($e) use ($self, $sharedManager) {
-            $self->onTriggerEvent($e, $sharedManager);
-        }, 100000000);
+
+        $this->listener = new TriggerEventListener($sharedManager);
     }
 
     /**
-     * @param EventInterface $e
-     * @param SharedEventManagerInterface $sharedManager
+     * @return TriggerEventListener
      */
-    public function onTriggerEvent(EventInterface $e, SharedEventManagerInterface $sharedManager)
+    public function getListener()
     {
-        $target = $e->getTarget();
-        $id = get_class($target);
-        $eventName = $e->getName();
-        $sharedEventMangerCallbacks = array();
-        $eventManagerCallbacks = array();
-        $calledTrace = 'Unknown';
-        $debugBacktrace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS);
-
-        if (isset($debugBacktrace[4]) && $debugBacktrace[4]['function'] === 'trigger') {
-            $calledTrace = $this->removeGetcwd($debugBacktrace[4]['file']) .':'.$debugBacktrace[4]['line'];
+        if (!$this->listener instanceof TriggerEventListener) {
+            throw new \BadMethodCallException('Method can not be called before init() method');
         }
-
-        if (is_object($target) && $target instanceof EventsCapableInterface) {
-            $em = $target->getEventManager();
-
-            if ($em instanceof EventManager) {
-                $listeners = $em->getListeners($eventName);
-                $eventManagerCallbacks = $this->getCallbacks($listeners);
-            }
-        }
-
-        $sharedListeners = $sharedManager->getListeners($id, $eventName);
-
-        if ($sharedListeners !== false) {
-            $sharedEventMangerCallbacks = $this->getCallbacks($sharedListeners);
-        }
-        $this->events[$this->getEventName($id, $eventName)] = array(
-            'Called in ' => $calledTrace,
-            'EventManager' => $eventManagerCallbacks,
-            'SharedEventManger' => $sharedEventMangerCallbacks,
-        );
-    }
-
-    /**
-     * @param string $id
-     * @param string $name
-     * @return string
-     */
-    protected function getEventName($id, $name)
-    {
-        return $id .'::'. $name;
-    }
-
-    /**
-     * @param PriorityQueue $listeners
-     * @return array[]
-     */
-    protected function getCallbacks(PriorityQueue $listeners)
-    {
-        $callbacks = array();
-
-        foreach ($listeners as $listener) {
-            if ($listener instanceof CallbackHandler) {
-                $callbacks[] = $this->getCallbackFromListener($listener);
-            }
-        }
-        return $callbacks;
-    }
-
-    /**
-     * @param CallbackHandler $listener
-     * @return array
-     */
-    protected function getCallbackFromListener(CallbackHandler $listener)
-    {
-        $callback = $listener->getCallback();
-        $priority = (int) $listener->getMetadatum('priority');
-
-        if ($callback instanceof \Closure) {
-            $ref = new \ReflectionFunction($callback);
-            $callbackId = 'Closure: '. $this->removeGetcwd($ref->getFileName()) .':'. $ref->getStartLine() .'-'. $ref->getEndLine();
-        } elseif (is_array($callback) && count($callback) === 2 && is_object($callback[0])) {
-            $callbackId = get_class($callback[0]) . '::' . $callback[1];
-        } elseif (is_string($callback)) {
-            $callbackId = $callback;
-        } else {
-            $callbackId = 'Unknown callback';
-        }
-        return array(
-            'callback' => $callbackId,
-            'priority' => $priority,
-        );
-    }
-
-    /**
-     * @param string $path
-     * @return string
-     */
-    protected function removeGetcwd($path)
-    {
-        $prefix = $this->getGetcwd();
-        if (substr($path, 0, strlen($prefix)) === $prefix) {
-            $path = substr($path, strlen($prefix));
-        }
-        return $path;
-    }
-
-    /**
-     * @return string
-     */
-    protected function getGetcwd()
-    {
-        if ($this->getcwd === null) {
-            $this->getcwd = getcwd() .'/';
-        }
-        return $this->getcwd;
-    }
-
-    /**
-     * @return array
-     */
-    public function getEvents()
-    {
-        return $this->events;
+        return $this->listener;
     }
 
     /**
